crimson/osd: explicitly fail on objecter-related objclass methods
The `cls_cxx_gather()` and `cls_cxx_get_gathered_data()` methods
of the Objclass API (the interface between a _Ceph Class_ plugin
and an OSD) are used to access the OSD's `Objecter` instance which
is used by the proxy tiering and _remote reads_. Currently, these
features are intentionally not supported in crimson. Therefore,
the patch switches the crimson's implementation of theese methods
to explicitly return `EOPNOTSUPP`.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>