mon: dissociate app from pools on pool remove
Problem:
Application is not dissociated from a pool when fs is deleted or when a
pool is removed from a fs. This causes the pool to be tied to the first
pool that it was associated with. The pool can't be added back to the
original fs neither can it be attached to other fs even if the fs is
deleted.

Solution:
This commit now ensures that the application is dissociated from the
pool when it is removed from a fs or when the entire fs is deleted.

Another solution is to recreate a new pool using the same devices and
then add it to the desired fs using a different name. Although, this
might work, it unnecessarily leaves garbage pool names and ids around.

Fixes: https://tracker.ceph.com/issues/57285
Signed-off-by: Milind Changire <mchangir@redhat.com>