common/LogEntry: specialize fmt::formatter<LogEntry>
so we can use the formatter defined for `LogEntry` in fmtlib v9.
in this new version of fmtlib, it is required to define a specialization
for the formatted type even when it comes to the types with an override of
operator<<(). since we already have an override for `LogEntry`, let's define
the specialization for `fmt::formatter<LogEntry>`.

this change should address the FTBFS when building with fmtlib v9.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>