dokan: switch to Dokany v2
The stable version of Dokany v2 has been released. It introduces
significant performance improvements, plus v1 is likely going
to be deprecated eventually.

As per this document[1], there are very few changes that we need to
make in order to use Dokany v2. We've covered those changes
in the previous commits.

That being said, we're now bumping the Dokany requirement to
2.0.5.1000 (latest).

[1] https://github.com/dokan-dev/dokany/wiki/Update-Dokan-1.1.0-application-to-Dokany-2.0.0

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>