crimson/block: fix the device size calculation.
In BlockSegmentManager, super block is updated with device size.
But a small amount of device capacity is reserved to store the
super block information and other tracking information.
Number of segments is calculated after discounting super block
size and tracking information size. This creates a mismatch
with the actual available size versus actual number of segments.

Update the available size after considering the reserved device
capacity and the number of segments and segment size.

Signed-off-by: Aravind Ramesh <aravind.ramesh@wdc.com>