doc/index.rst - remove dboard as TOC top-level
This PR removes mgr/dashboard from the list of top-
level TOC entries in doc/index.rst.

Before this PR, the Dashboard documentation appeared
in the mgr documentation (and it still does), and it
also appeared as its own guide as a top-level entry in
the sidebar on the left of the rendered Sphinx documen-
tation.

This led to a curious feeling of deja vu for me as I worked
on the documentation, and I think that the documentation
suite will be clearer to the first-time reader without this
repetition. (I think it seems like it's two different documents,
making a reader feel that they must read both before deciding
which one to use. I can only imagine the irritation at their
realization that they have wasted their time comparing two bodies
of text that turned out to be identical.)

Signed-off-by: Zac Dover <zac.dover@gmail.com>