mgr/nfs: reformat long function calls to one line per arg
Using a hand rolled version of the long function call style similar to
that produced by the "Black" tool, put every argument to the
create_export function on a line of its own. This makes the code easier
to read and, importantly for an upcoming change, easier to add new
arguments.

Signed-off-by: John Mulligan <jmulligan@redhat.com>