mgr/nfs: add --sectype flag to nfs export create functions
Fixes: https://tracker.ceph.com/issues/57404

Add the sectype argument, also for use as a CLI option, to the functions
that create nfs exports. NFS Ganesha can use the sectype configuration
option to decide what kind of security will be applied/required for
connection to an export.

NOTE: This option is only useful when Ganesha is configured along with
LDAP/Kerberos integration. Configuration of that integration is outside
the scope of these patches.

Signed-off-by: John Mulligan <jmulligan@redhat.com>
(cherry picked from commit 11affedea600360afd1ee7e3bb62c51f7b423a67)