mgr/dashboard: fix Expected to find element: `cd-modal .badge but never found it
This looks related to https://github.com/ceph/ceph/pull/46323, but i am
not sure why it started failing recently in the teuthology.

The error is fixed by deselecting the row manually, so that it won't
accidentally get deselected when we actually want it to select it.

Fixes: https://tracker.ceph.com/issues/57207
Signed-off-by: Nizamudeen A <nia@redhat.com>
(cherry picked from commit dcbf179106273b85361dd6067239c4e826d1f3bb)