os/bluestore: Provide backup and restore for bluestore db
If have a function what we can backup rocksdb of bluetore,
and then we can use backup db to restore db when db damage.
At now, osd block device disk is a 16TB HDD, if have one
of db happend damage, we have to destroyed and redeployed osd,
and then have a 16TB data need to backfill, but block device
is intact. If we can restore db by backup, we don't need to
redeployed osd, just backfill a little data.

Signed-off-by: Yite Gu <ess_gyt@qq.com>