osd: Reduce default max backfill/recoveries for mclock to realistic limits
The earlier max number of backfill and recovery limits were overly
optimistic and caused recovery ops to overwhelm client ops regardless of
the mclock profile type.

Set the max backfill limit to 10. Set the max recovery limit to 10 for
osds having HDD backing device. For osds having SSD backing device, set
the recovery limit to 20.

Introduce and new config option, 'osd_mclock_override_recovery_settings',
that when set (false by default) allows the modification of the max
backfill and recovery options with mclock scheduler enabled.

It is important to note that even with the
'osd_mclock_override_recovery_settings' disabled, the config subsystem
reflects the changes made to the recovery and backfill settings when
using the cli commands. But the osd in this case reverts the recovery
and backfill limits to the mClock defaults and raises a cluster log
warning indicating the revert. This should force the user to set the
'osd_mclock_override_recovery_settings' option first and then make
changes to the recovery and backfill limits.

Fixes: https://tracker.ceph.com/issues/57529
Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>
(cherry picked from commit 16f6ca1c5972d68385c5bfed777bbed0f3aceaf7)