rgw: rename request traces and change for tags
when we have a lot of traces, for example on multisite configuration
we have traces for list_meta_logs and list_data_changes and
it become too verbose to have trans id as the trace name,
the filtering in the jaeger gui becomes a mess,
so we will have the trans_id as tag, so ops of same
type will be named the op name only for cleaner usage.

the tag 'return' is ambigious. the actual name should be 'op_result'
so we can distinguish between http return code and op return value.

the tag 'type' doesn't add extra information, so it removed.

Signed-off-by: Omri Zeneva <ozeneva@redhat.com>