mgr/cephadm: iscsi username and password defaults to admin
A while ago, the iscsi username and password were defaulted to admin the
config_dashboard part. But it was not saved to the
`iscsi-gateway.cfg.j2` which results in a misconfiguration.

For eg, if you deploy an iscsi service with service_spec which doesn't
have the api_user and api_password field it results the dashboard to
show the gateways as always down because dashboard was unable to make
request to the iscsi server and every api calls result in a `[dashboard
ERROR rest_client] iscsi REST API failed GET req status: 401`.

Fixes: https://tracker.ceph.com/issues/57730
Signed-off-by: Nizamudeen A <nia@redhat.com>
(cherry picked from commit bf87bc797ff0b47a60f7169698bb8ed2a5080c14)