doc/glossary.rst: remove duplicates
This commit removes similar but distinct entries for the following:
   * CephFS
   * Ceph Client

Removal of a glossary term that is referred to in the body of the
documentation suite requires the alteration of the text string
that refers to the glossary term. Alterations of this kind have
been made to doc/architecture.rst and doc/rados/api/index.rst.

(cherry picked from commit 0692b223540143cdb4a60b54595a7069284b5481)
Signed-off-by: Zac Dover <zac.dover@gmail.com>